feat: move input limit check to verify_inputs
